@@ -1,38 +1,35 @@
 {
+    "type": "module",
     "private": true,
     "scripts": {
-        "development": "mix",
-        "dev": "npm run development",
-        "watch": "mix watch",
-        "serve": "npm run watch",
-        "watch-poll": "mix watch -- --watch-options-poll=1000",
-        "hot": "mix watch --hot",
-        "prod": "npm run production",
-        "build": "npm run production",
-        "production": "mix --production"
+        "dev": "vite",
+        "build": "vite build",
+        "serve": "vite preview"
     },
     "devDependencies": {
-        "axios": "^0.21.1",
-        "bootstrap": "^4.5.3",
-        "bootstrap-datepicker": "^1.8.0",
-        "bootstrap-vue": "^2.18.0",
-        "clean-webpack-plugin": "^3.0.0",
-        "cross-env": "^5.1",
+        "@vitejs/plugin-vue": "^5.0.4",
+        "@vue/compiler-sfc": "^3.4.27",
+        "axios": "^1.6.8",
+        "bootstrap-datepicker": "^1.10.0",
+        "bootstrap-vue": "^2.23.1",
+        "clean-webpack-plugin": "^4.0.0",
+        "cross-env": "^7.0.3",
         "jquery": "^3.5.1",
+        "laravel-vite-plugin": "^1.2.0",
         "lodash": "^4.17.21",
         "moment": "^2.29.1",
         "popper.js": "^1.12",
-        "resolve-url-loader": "^3.1.2",
-        "sass": "^1.27.0",
-        "sass-loader": "^8.0.2",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.11.1",
-        "vue-router": "^3.4.7",
-        "vue-template-compiler": "^2.6.12",
+        "resolve-url-loader": "^5.0.0",
+        "sass": "^1.86.0",
+        "sass-loader": "^16.0.5",
+        "vite": "^6.2.3",
+        "vue-loader": "^17.4.2",
+        "vue-router": "^4.5.0",
         "vuejs-datepicker": "^1.2.0",
-        "vuex": "^3.5.1"
+        "vuex": "^4.1.0"
     },
     "dependencies": {
-        "laravel-mix": "^6.0.43"
+        "@vue/compat": "3.5.13",
+        "vue": "^3.4.27"
     }
 }
